dont do for non full slice
